refactor(decomp): hoist RBAC context types to proximadb-tenant (Slice D link)
Hoist the four leaf RBAC authorization-context data types out of the root
`src/security/rbac_service.rs` into the foundation `proximadb-tenant` crate,
unblocking the engine-port-traits module move (the types that
`src/storage/traits/query.rs` imports from rbac_service are now crate-accessible).

Types moved (all leaf data — no root-internal deps, only String/chrono/serde/uuid/std):
  - UnifiedPermission (enum, ~40 variants)
  - UnifiedAuthMethod (enum)
  - UnifiedUserContext (struct) + its 4 inherent methods (anonymous /
    is_authenticated / is_session_expired / display_name — moved from
    auth_service.rs since inherent impls must live in the defining crate)
  - RbacTenantContext (struct)

Source-compatible: rbac_service.rs keeps `pub use proximadb_tenant::<Type>` re-exports
for every moved type, plus the existing `TenantContext`/`AuthMethod` aliases, so all
existing crate::security::rbac_service::<Type> call sites (~477 refs across the 4 types)
resolve unchanged. The data-type unit test (test_unified_permission_variants) moved with
UnifiedPermission to the tenant crate; the manager-behavior tests stay in rbac_service.rs
and pass via the re-export shim.

Verified: cargo check --lib --bins, --tests -p proximadb/-tenant, nextest -p proximadb-tenant
(17/17), nextest rbac lib tests (33/33), clippy --lib --bins, fmt, workspace-boundaries.
