+++ crates/foundation/proximadb-tenant/Cargo.toml
@@ -16,8 +16,10 @@ path = "src/lib.rs"
 [dependencies]
 proximadb-proto.workspace = true
 serde.workspace = true
-anyhow.workspace = true
 serde_json.workspace = true
+chrono.workspace = true
+uuid = { version = "1.0", features = ["v4"] }
+anyhow.workspace = true
 async-trait.workspace = true
 tokio.workspace = true
 tracing.workspace = true
